Projeto_Integrador/projeto_integrador_7/MQTT_Flask/webapp/requirements.txt
Flask==2.0.1
Flask-Admin==1.5.8
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-Migrate==3.0.0
Flask-MQTT==1.1.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.0
Jinja2==3.0.1
passlib==1.7.4
psycopg2-binary==2.8.6
python-dotenv==0.17.1
requests==2.25.1
SQLAlchemy==1.4.15
WTForms==2.3.3
